+++ src/test/io_test.py
@@ -9,6 +9,7 @@ from olmo_core.io import (
     file_exists,
     glob_directory,
     list_directory,
+    remove_file,
     serialize_to_tensor,
     upload,
 )
@@ -95,6 +96,13 @@ def _run_remote_functionality(tmp_path, remote_dir):
     # Copy dir.
     copy_dir(f"{remote_dir}", tmp_path / "dir3")
     assert (tmp_path / "dir3/dir1/file2.json").is_file()
+
+    # Remove a file from the remote dir.
+    remove_file(f"{remote_dir}/file1.json")
+    assert set(list_directory(remote_dir, recurse=True)) == {
+        f"{remote_dir}/dir1",
+        f"{remote_dir}/dir1/file2",
+    }
 
 
 def test_s3_functionality(tmp_path, s3_checkpoint_dir):